project: added custom container implementation
Problems with docker, runc and nspawn:
- Docker and runc are not very stable either API-wise, nor in runtime
- Mountpoints mounted after starting the container do not propagate inside,
  because docker and runc both set the propagation to PRIVATE instead of
  SLAVE
- Requiring people to have nspawn sounded unreasonable
- These all do way more than we actually need, and we have to disable
  many features manually (e.g. capability settings, network namespaces)

What we have to do to create a container is basically:
- Create new namespaces with clone and unshare
- In the new mount namespace, pivot_root to the desired root directory
- Mount the default directories needed by programs (/proc, /sys, etc.)
- Create the essential device nodes (null, zero, randm), but not actually
  mount /dev
- Umount the old root directory
- Reap zombies with PID1
- Wait for an exit signal (eof on the control pipe)

This is all we need for almost complete containment. Of course we don't
limit capabilities or network namespaces, as this is not for security,
only to protect /dev and /proc.

An interesting note is that the clone syscall was used instead of the clone
command, as the latter does way more than we actually need here, and is
actually designed to start threads. We just needed a fork with the pid
namespace flag.

Execing other processes in the container is done via the 'nsenter' command,
which is ubiquitous, and does the hard part of PID namespaces ("fork();
nsenter(); fork(); waitpid()"), and is very fast. The only overhead it has
over a pure python implementation is an extra exec and a few open() calls.

References: jira PL-188
Change-Id: I77f91e2c87e7668279f5e1a687d91e0c5bdb6bfd