bake/container: set hostname to localhost
/etc/hosts file has a static content with entries such as
"127.0.0.1 localhost". The "hostname --fqdn" command looks up the hostname
via gethostname(), and calls getaddrinfo on it, which in turn parses the
hosts file. If the current hostname is not in the hosts file, this
operation will fail.

So we set the hostname to "localhost" in the container namespace, and just
to avoid confusion, in /etc/hostname too.

Change-Id: I477ed2107abdd36acbea1157fd56114c0ee6d326