container: optionally isolate networking
Most of our building containers don't actually need outside networking,
and if they e.g. start a daemon, that could interfere with the host
computer.
We thus create an unshared network namespace by default and disable this
functionality for the few builders that actually need networking (such
as packageinstall).
A room for improvement here is that only the local apt proxy needs to be
available inside the containers, so we could make only that available
inside, not the whole host net ns.

Change-Id: Ia4270a149f61486ba093621c820f7c00f05e0134
References: jira PL-298