container: also create loop device nodes
To allow losetup, mount -o loop, and kpartx to work. These are needed
for e.g. the diskimagebuilder.
This is not ideal, because the hosts loop devices will leak into the
container, but this is way quicker now than rewriting diskimagebuilder
in python (at least until the partitioning and mounting those
partitions with kpartx)

Change-Id: I479fa0183f4ae6bc74be4525f424cf515389ecc3