container/context.py: interactive_shell: do not use bashrc from the firmware
When there is a bashrc on the firmware that prints error messages when started in bake,
it can confuse the bake user, who might think that error is related to the build error.

This patch runs bash without starting bashrc and bash_profile, and uses a custom-generated
PS1 so the user knows which node build failed.

Change-Id: I484356e2c376511ca8e31d42935a34c2ac8acac8