pid1.py: use umount2 with MNT_DETACH to clean /old_root
After pivot_root'ing to the container directory, all entries
must be unmounted in /old_root, giving us a host-independent
environment inside the container.

Previously a separate call was used to determine what
directories need to be unmounted and the directories were
iterated over with an umount call. However, there is a more
efficient way to do this, using the MNT_DETACH flag of
umount2, which will return immediately, while the kernel
does exactly the same job in the background.

Change-Id: I4f752a0ddcec1e1a5a9b52d659e70b66eb4980b3