container/context.py: avoid calling clone() directly
It turned out that forking (or cloning) from a multi-threaded python
program has some dangers, which is correctly handled in os.fork() and
subprocess module, but not handled in our own naked clone() implementation.

If clone() is called without the python wrapping (like in os.fork()),
locks and other thread-related memory structures are copied to the new
process. Since there is a single thread in the forked process, it is
possible that e.g. GIL was locked by a thread in the main process and it
is never released in the child, locking up its execution.

To avoid this, the new pid ns is created via unshare, before starting
the new process, and os.fork() is used for creating a new process, which
takes care about the locks in threads.

Change-Id: I1ba492f43dff96231818cb32238c6c7a57194a80