furnace/context.py: bind mount '/etc/resolv.conf' in host network mode
With this solution the '/etc/resolv.conf' of the host can be used inside
the container, because it is bind mounted before every run in host network
mode (when network is not isolated). The bind mount is read-only to avoid
the modification of the '/etc/resolv.conf' of the host machine. With this
commit bind mounts can be added to the ContainerContext.