pid1: map all loop devices
Up to now furnace created 8 loop devices in /dev inside the container.
This behavior caused problems for those, who used 8 or more loop devices
before the container started. The solution is that pid1 creates as many
loop devices as currently exist in the system.

The _get_loop_devices() method creates a generator of loop devices.
Each loop device contains a name and a device minor number.

Change-Id: I2ca4a011c377d89d7f71bda261858036a8b5e495