test: fix tests and make them able to run on multiple OS
On the test machine the PATH variable may not set properly, that is why
I modified programs' paths to an absolute form.
The tests use an Ubuntu Xenial root filesystem, where the crucial
programs live under /bin or /sbin. On Archlinux there are no programs
under /bin or /sbin (these are only symlinks to the corresponding
directories under /usr) so the PATH variable does not contain these
directories.

The pytest.raises message parameter was removed in pytest 5.0.
Learn more:
https://docs.pytest.org/en/5.1.2/deprecations.html#raises-message-deprecated