.github/workflows/release.yml: add release process
Github actions cannot push a protected branch, that is why the release
process splited two parts.
The first part is the 'version bump' workflow, bumps the version and
create a pull a request. After that an administrator should perform
a 'rebase and merge' for this pull request.
The second part is the 'release' workflow, creates a wheel, performs
a relase testing, publishes the package to PYPI, and creates a tag
from the actual version.